+++ .vscode/settings.json
@@ -27,6 +27,7 @@
     "Flicking",
     "fontaine",
     "found",
+    "ghost",
     "headline",
     "hello",
     "icons",
@@ -52,6 +53,7 @@
     "recommended",
     "Rights",
     "sitemap",
+    "Slideover",
     "slider",
     "slides",
     "social",
