+++ assets/scss/basic/fonts.scss
@@ -3,8 +3,66 @@
 */
 @font-face {
   font-family: Pretendard;
+  src: local('~/assets/fonts/Pretendard-Thin.otf');
+  font-weight: 100;
+  font-style: normal;
   font-display: swap;
-  src: url('https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.8/dist/web/static/pretendard.css') format('woff2');
+}
+@font-face {
+  font-family: Pretendard;
+  src: local('~/assets/fonts/Pretendard-ExtraLight.otf');
+  font-weight: 200;
+  font-style: normal;
+  font-display: swap;
+}
+@font-face {
+  font-family: Pretendard;
+  src: local('~/assets/fonts/Pretendard-Light.otf');
+  font-weight: 300;
+  font-style: normal;
+  font-display: swap;
+}
+@font-face {
+  font-family: Pretendard;
+  src: local('~/assets/fonts/Pretendard-Regular.otf');
+  font-weight: 400;
+  font-style: normal;
+  font-display: swap;
+}
+@font-face {
+  font-family: Pretendard;
+  src: local('~/assets/fonts/Pretendard-Medium.otf');
+  font-weight: 500;
+  font-style: normal;
+  font-display: swap;
+}
+@font-face {
+  font-family: Pretendard;
+  src: local('~/assets/fonts/Pretendard-SemiBold.otf');
+  font-weight: 600;
+  font-style: normal;
+  font-display: swap;
+}
+@font-face {
+  font-family: Pretendard;
+  src: local('~/assets/fonts/Pretendard-Bold.otf');
+  font-weight: 700;
+  font-style: normal;
+  font-display: swap;
+}
+@font-face {
+  font-family: Pretendard;
+  src: local('~/assets/fonts/Pretendard-ExtraBold.otf');
+  font-weight: 800;
+  font-style: normal;
+  font-display: swap;
+}
+@font-face {
+  font-family: Pretendard;
+  src: local('~/assets/fonts/Pretendard-Black.otf');
+  font-weight: 900;
+  font-style: normal;
+  font-display: swap;
 }
 @font-face {
   font-family: Roboto;
