+++ package.json
@@ -65,9 +65,6 @@
     "@tiptap/pm": "^2.1.12",
     "@tiptap/starter-kit": "^2.1.12",
     "@tiptap/vue-3": "^2.1.12",
-    "@tresjs/cientos": "^3.7.0",
-    "@tresjs/nuxt": "^2.0.1",
-    "@types/three": "^0.160.0",
     "@typescript-eslint/parser": "6.15.0",
     "@vercel/analytics": "^1.1.1",
     "@vercel/speed-insights": "1.0.2",
@@ -89,7 +86,6 @@
     "stylelint-config-recommended-vue": "^1.5.0",
     "stylelint-config-standard-scss": "12.0.0",
     "supabase": "1.127.1",
-    "three": "^0.160.0",
     "typeit": "^8.8.0",
     "typescript": "5.3.3",
     "vue-dompurify-html": "5.0.1"
